restructureed with tinsy bit of style
